requirements.txt
install/modem-1.1-py3-none-any.whl
pyelftools==0.27
pyserial==3.5

